requirements.txt
numpy==1.24.3
pygame==2.5.0
neat-python==0.92
matplotlib==3.7.1 